Move Stream Functions to a Directory
